Add VassalCard (#28)
Adds Vassal.

One interesting decision here is how to handle the choice to play the discarded card if it's an action. I chose to first discard it, then give the player an option to play it from the top of the discard pile. For the base set this should be fine but not sure how it may interact with future cards. For example, if some ability was triggered on discarding, then it may not end up in the discard pile at all. Then this logic might try to play a different card (the one that did end up on the top of the discard pile).